+++ selfdrive/ui/SConscript
@@ -30,7 +30,7 @@ qt_src = ["main.cc", "qt/sidebar.cc", "qt/body.cc",
           "qt/window.cc", "qt/home.cc", "qt/offroad/settings.cc",
           "qt/offroad/software_settings.cc", "qt/offroad/onboarding.cc",
           "qt/offroad/driverview.cc", "qt/offroad/experimental_mode.cc",
-          "qt/offroad/toggles_panel.cc",
+          "qt/offroad/toggles_panel.cc", "qt/offroad/device_panel.cc",
           "qt/onroad/onroad_home.cc", "qt/onroad/annotated_camera.cc",
           "qt/onroad/buttons.cc", "qt/onroad/alerts.cc",]
