+++ selfdrive/ui/qt/offroad/settings.h
@@ -35,26 +35,6 @@ private:
   QStackedWidget *panel_widget;
 };
 
-class DevicePanel : public ListWidget {
-  Q_OBJECT
-public:
-  explicit DevicePanel(SettingsWindow *parent);
-  void showEvent(QShowEvent *event) override;
-
-signals:
-  void reviewTrainingGuide();
-  void showDriverView();
-
-private slots:
-  void poweroff();
-  void reboot();
-  void updateCalibDescription();
-
-private:
-  Params params;
-  ButtonControl *pair_device;
-};
-
 class SoftwarePanel : public ListWidget {
   Q_OBJECT
 public:
